refactor(iot-checkin): extract checkout-gate policy into CheckinService
Drive the three assigned handlers under the gocognit-15 ratchet by moving
the daily-checkout / checkout-time-gate policy into services/iot/checkin and
decomposing the check-in and pickup-query handlers into small shells.

- Move the daily-checkout gate policy (ResolveRawDailyCheckoutTime,
  studentDailyCheckoutTime, IsAfterCheckoutTimeGate, ShouldShowDailyCheckoutWithGroup,
  ShouldUpgradeToDailyCheckout, and the timeNow clock hook) from the api
  handler package into CheckinService, next to the existing check-in decision
  flow. IsAfterCheckoutTimeGate is split into per-student and global sub-gates.
- Add ResolveActiveStudentCount to CheckinService for the device-scoped
  active-student heartbeat/count orchestration.
- Wire Settings/Pickup/Education into CheckinService (factory + deps); drop the
  now-orphaned EducationService field from the api checkin Resource.
- Decompose deviceCheckin (33->13) and devicePickupQuery (21->~9) into
  presentation/enrichment helpers; the pickup-query error contract (500 on
  lookup failure vs 404/400 classification) is preserved byte-for-byte.
- Relocate the policy tests to services/iot/checkin internal tests; api-layer
  response-shaping tests stay. Remove the three allowlist entries.

Wire strings and response shapes unchanged (PyrePortal contract).
