+++ backend/test/handler_complexity_ratchet_test.go
@@ -57,9 +57,6 @@ var handlerComplexityAllowlist = map[string]int{
 	"api/enrollment/schema_handlers.go:(*Resource).updateSchema":                                19,
 	"api/guardians/handlers.go:(*Resource).deleteGuardian":                                      22,
 	"api/iot/checkin/attendance_handlers.go:(*AttendanceResource).handleDailyCheckout":          18,
-	"api/iot/checkin/handlers.go:(*Resource).deviceCheckin":                                     33,
-	"api/iot/checkin/handlers.go:(*Resource).devicePickupQuery":                                 21,
-	"api/iot/checkin/helpers.go:(*Resource).isAfterCheckoutTimeGate":                            17,
 	"api/operator/api.go:(*Resource).Router":                                                    20,
 	"api/operator/settings.go:(*SettingsResource).ResetSchoolSettingValue":                      17,
 	"api/operator/settings.go:(*SettingsResource).SetSchoolSettingValue":                        19,
